MB1-143 Refactor: 초기 사용자 설정 등록하는 useMutation을 커스텀훅으로 분리
